functions/requirements.txt
firebase-admin
google-auth-oauthlib
google-api-python-client
requests
firebase-functions